requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.29.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
prometheus-client==0.20.0
httpx==0.27.2
